Created Math library functions, bugfix

+++ inc/core/math.h
@@ -1,6 +1,8 @@
 #ifndef MATH_H
 #define MATH_H
 
+#include <stdexcept>
+
 class Math
 {
 public:
@@ -8,9 +10,9 @@ public:
     static double sub(double a, double b);
     static double mul(double a, double b);
     static double div(double a, double b);
-    static double exp(double x, unsigned int exp);
+    static double exp(double x, int exp);
     static double sqrt(double x);
-    static double fact(double x);
+    static int fact(int x);
 };
 
 #endif // MATH_H
